agent/requirements.txt
# Core dependencies
pydantic
function-stream

# Google ADK and GenAI dependencies
google-adk
google-genai

# Apache Pulsar client
pulsar-client

# JSON processing
jsonnet>=0.20.0

# Type hints support
typing-extensions

# Database dependencies
asyncpg
sqlalchemy[asyncio]

# Testing dependencies
pytest
pytest-asyncio